Add "contact with source case" to epi data #629
Add "4-Was in close proximity (1 meter) with source case" to contact proximity types (former 4 becomes 5)
add an "Resulting case assigned by" field to the contact that automatically persists who assigned a resulting case to a contact. If the case was automatically assigned by the system the field stays empty.
replace case entity reference in app contact with uuid of case so the case does not have to be transferred
add disease field to contact dto, so the source case does not have to be transfered