Cleanup files on disk
Documents are deleted "immediately" (at the end of the
transaction). This means that backups should be "atomic"
across the database and filesystem.

An alternative implementation would be to mark the
documents as deleted in the database, and use a cron job
to actually delete both the file on disk and document in
the database.
This doesn't apply to the documents that failed to be
saved in the database due to a transaction rollback,
where the file on disk is guaranteed to never have been
referenced in the database.

Change-Id: If6ea21d17f1336440e43679b2d68e6e092c74c0e